+++ billy/web/public/views/committees.py
@@ -9,7 +9,6 @@ from django.views.decorators.csrf import ensure_csrf_cookie
 from djpjax import pjax
 
 from billy.core import settings
-from billy.utils import popularity
 from billy.models import db, Metadata, DoesNotExist
 
 from .utils import templatename, mongo_fields
@@ -102,8 +101,6 @@ def committee(request, abbr, committee_id):
     committee = db.committees.find_one({'_id': committee_id})
     if committee is None:
         raise Http404
-
-    popularity.counter.inc('committees', committee_id, abbr=abbr)
 
     return render(request, templatename('committee'),
                   dict(committee=committee, abbr=abbr,